BS-13403 In Form Mapping, pencil icon should generate the form based on the contract and do the mapping (same as UI Designer icon in contract view)
move code from formMappingComposite to formReferenceExpressionViewer
