+++ bundles/plugins/org.bonitasoft.studio.designer/src/org/bonitasoft/studio/designer/ui/property/section/control/FormReferenceExpressionViewer.java
@@ -15,14 +15,28 @@
 package org.bonitasoft.studio.designer.ui.property.section.control;
 
 import org.bonitasoft.studio.common.ExpressionConstants;
+import org.bonitasoft.studio.common.emf.tools.ExpressionHelper;
+import org.bonitasoft.studio.common.repository.RepositoryAccessor;
+import org.bonitasoft.studio.designer.core.command.UpdateFormMappingCommand;
+import org.bonitasoft.studio.designer.core.expression.CreateNewFormProposalListener;
 import org.bonitasoft.studio.designer.core.repository.WebPageFileStore;
 import org.bonitasoft.studio.designer.core.repository.WebPageRepositoryStore;
 import org.bonitasoft.studio.expression.editor.viewer.EditExpressionDialog;
 import org.bonitasoft.studio.expression.editor.viewer.ExpressionViewer;
+import org.bonitasoft.studio.model.expression.Expression;
 import org.bonitasoft.studio.model.expression.ExpressionPackage;
+import org.bonitasoft.studio.model.process.FormMapping;
 import org.eclipse.core.databinding.conversion.Converter;
+import org.eclipse.core.databinding.observable.value.IObservableValue;
 import org.eclipse.emf.ecore.EObject;
+import org.eclipse.emf.transaction.TransactionalEditingDomain;
+import org.eclipse.emf.transaction.util.TransactionUtil;
+import org.eclipse.swt.SWT;
+import org.eclipse.swt.events.SelectionAdapter;
+import org.eclipse.swt.events.SelectionEvent;
 import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Listener;
+import org.eclipse.swt.widgets.ToolItem;
 import org.eclipse.ui.views.properties.tabbed.TabbedPropertySheetWidgetFactory;
 
 /**
@@ -81,6 +95,62 @@ public class FormReferenceExpressionViewer extends ExpressionViewer {
 
     protected void updateName(final String newName) {
         expressionItemProvider.setPropertyValue(getSelectedExpression(), ExpressionPackage.Literals.EXPRESSION__NAME.getName(), newName);
+    }
+
+    /**
+     * @param formMappingObservable
+     */
+    public void setEditControlBehavior(final IObservableValue formMappingObservable, final CreateNewFormProposalListener createNewFormListener,
+            final RepositoryAccessor repositoryAccessor) {
+        final ToolItem editControl = getButtonControl();
+        clearExistingSelectionListeners(editControl);
+        editControl.addSelectionListener(new SelectionAdapter() {
+
+            @Override
+            public void widgetSelected(final SelectionEvent e) {
+                createOReditForm(formMappingObservable, createNewFormListener, repositoryAccessor);
+            }
+
+        });
+    }
+
+    private void clearExistingSelectionListeners(final ToolItem editControl) {
+        final Listener[] toRemove = editControl.getListeners(SWT.Selection);
+        for (final Listener l : toRemove) {
+            editControl.removeListener(SWT.Selection, l);
+        }
+    }
+
+    protected void createOReditForm(final IObservableValue formMappingObservable, final CreateNewFormProposalListener createNewFormListener,
+            final RepositoryAccessor repositoryAccessor) {
+        final FormMapping mapping = (FormMapping) formMappingObservable.getValue();
+        final Expression targetForm = mapping.getTargetForm();
+        if (targetForm.hasContent()) {
+            repositoryAccessor.getRepositoryStore(WebPageRepositoryStore.class).getChild(targetForm.getContent()).open();
+        } else {
+            createNewForm(mapping, createNewFormListener, repositoryAccessor);
+        }
+    }
+
+    protected void createNewForm(final FormMapping mapping, final CreateNewFormProposalListener createNewFormListener,
+            final RepositoryAccessor repositoryAccessor) {
+        final String newPageId = createNewFormListener.handleEvent(mapping, null);
+        final WebPageRepositoryStore repositoryStore = repositoryAccessor.getRepositoryStore(WebPageRepositoryStore.class);
+        repositoryStore.refresh();
+        final WebPageFileStore webPageFileStore = repositoryStore.getChild(newPageId);
+        if (webPageFileStore != null) {
+            final TransactionalEditingDomain editingDomain = getEditingDomain(mapping);
+            editingDomain.getCommandStack().execute(new UpdateFormMappingCommand(editingDomain, mapping,
+                    ExpressionHelper.createFormReferenceExpression(webPageFileStore.getDisplayName(), newPageId)));
+        }
+    }
+
+    /**
+     * @param mapping
+     * @return
+     */
+    public TransactionalEditingDomain getEditingDomain(final FormMapping mapping) {
+        return TransactionUtil.getEditingDomain(mapping.eContainer());
     }
 
 }
