+++ bundles/plugins/org.bonitasoft.studio.designer/src/org/bonitasoft/studio/designer/ui/property/section/control/InternalMappingComposite.java
@@ -15,21 +15,17 @@
 package org.bonitasoft.studio.designer.ui.property.section.control;
 
 import org.bonitasoft.studio.common.ExpressionConstants;
-import org.bonitasoft.studio.common.emf.tools.ExpressionHelper;
 import org.bonitasoft.studio.common.jface.databinding.CustomEMFEditObservables;
 import org.bonitasoft.studio.common.repository.RepositoryAccessor;
-import org.bonitasoft.studio.designer.core.command.UpdateFormMappingCommand;
 import org.bonitasoft.studio.designer.core.expression.CreateNewFormProposalListener;
 import org.bonitasoft.studio.designer.core.repository.WebPageFileStore;
 import org.bonitasoft.studio.designer.core.repository.WebPageRepositoryStore;
 import org.bonitasoft.studio.designer.i18n.Messages;
 import org.bonitasoft.studio.designer.ui.property.section.FormReferenceProposalLabelProvider;
 import org.bonitasoft.studio.expression.editor.filter.AvailableExpressionTypeFilter;
-import org.bonitasoft.studio.expression.editor.viewer.ExpressionViewer;
 import org.bonitasoft.studio.model.expression.Expression;
 import org.bonitasoft.studio.model.expression.provider.ExpressionItemProvider;
 import org.bonitasoft.studio.model.expression.provider.ExpressionItemProviderAdapterFactory;
-import org.bonitasoft.studio.model.process.FormMapping;
 import org.bonitasoft.studio.model.process.ProcessPackage;
 import org.bonitasoft.studio.preferences.BonitaPreferenceConstants;
 import org.eclipse.core.databinding.DataBindingContext;
@@ -37,8 +33,6 @@ import org.eclipse.core.databinding.observable.Realm;
 import org.eclipse.core.databinding.observable.value.IObservableValue;
 import org.eclipse.core.resources.IResourceChangeListener;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences;
-import org.eclipse.emf.transaction.TransactionalEditingDomain;
-import org.eclipse.emf.transaction.util.TransactionUtil;
 import org.eclipse.jface.databinding.viewers.ViewersObservables;
 import org.eclipse.jface.layout.GridDataFactory;
 import org.eclipse.jface.layout.GridLayoutFactory;
@@ -46,12 +40,8 @@ import org.eclipse.jface.viewers.ISelectionChangedListener;
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.viewers.SelectionChangedEvent;
 import org.eclipse.swt.SWT;
-import org.eclipse.swt.events.SelectionAdapter;
-import org.eclipse.swt.events.SelectionEvent;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Label;
-import org.eclipse.swt.widgets.Listener;
-import org.eclipse.swt.widgets.ToolItem;
 import org.eclipse.ui.views.properties.tabbed.TabbedPropertySheetWidgetFactory;
 
 /**
@@ -61,7 +51,7 @@ public class InternalMappingComposite extends Composite implements BonitaPrefere
 
     private static final int WIDTH_HINT = 300;
 
-    private final ExpressionViewer targetFormExpressionViewer;
+    private final FormReferenceExpressionViewer targetFormExpressionViewer;
     private final RepositoryAccessor repositoryAccessor;
     private final WebPageNameResourceChangeListener webPageNameResourceChangeListener;
     private final CreateNewFormProposalListener createNewFormListener;
@@ -131,53 +121,7 @@ public class InternalMappingComposite extends Composite implements BonitaPrefere
         context.bindValue(ViewersObservables.observeInput(targetFormExpressionViewer), formMappingObservable);
         context.bindValue(ViewersObservables.observeSingleSelection(targetFormExpressionViewer),
                 CustomEMFEditObservables.observeDetailValue(Realm.getDefault(), formMappingObservable, ProcessPackage.Literals.FORM_MAPPING__TARGET_FORM));
-
-        final ToolItem editControl = targetFormExpressionViewer.getButtonControl();
-        clearExistingSelectionListeners(editControl);
-        editControl.addSelectionListener(new SelectionAdapter() {
-
-            @Override
-            public void widgetSelected(final SelectionEvent e) {
-                createOReditForm(formMappingObservable);
-            }
-
-        });
+        targetFormExpressionViewer.setEditControlBehavior(formMappingObservable, createNewFormListener, repositoryAccessor);
     }
 
-    private void clearExistingSelectionListeners(final ToolItem editControl) {
-        final Listener[] toRemove = editControl.getListeners(SWT.Selection);
-        for (final Listener l : toRemove) {
-            editControl.removeListener(SWT.Selection, l);
-        }
-    }
-
-    protected void createOReditForm(final IObservableValue formMappingObservable) {
-        final FormMapping mapping = (FormMapping) formMappingObservable.getValue();
-        final Expression targetForm = mapping.getTargetForm();
-        if (targetForm.hasContent()) {
-            repositoryAccessor.getRepositoryStore(WebPageRepositoryStore.class).getChild(targetForm.getContent()).open();
-        } else {
-            createNewForm(mapping);
-        }
-    }
-
-    protected void createNewForm(final FormMapping mapping) {
-        final String newPageId = createNewFormListener.handleEvent(mapping, null);
-        final WebPageRepositoryStore repositoryStore = repositoryAccessor.getRepositoryStore(WebPageRepositoryStore.class);
-        repositoryStore.refresh();
-        final WebPageFileStore webPageFileStore = repositoryStore.getChild(newPageId);
-        if (webPageFileStore != null) {
-            final TransactionalEditingDomain editingDomain = getEditingDomain(mapping);
-            editingDomain.getCommandStack().execute(new UpdateFormMappingCommand(editingDomain, mapping,
-                    ExpressionHelper.createFormReferenceExpression(webPageFileStore.getDisplayName(), newPageId)));
-        }
-    }
-
-    /**
-     * @param mapping
-     * @return
-     */
-    public TransactionalEditingDomain getEditingDomain(final FormMapping mapping) {
-        return TransactionUtil.getEditingDomain(mapping.eContainer());
-    }
 }
